Moved Differ and Difference from org.scalatest to org.scalactic.
