+++ scalactic/src/main/scala/org/scalactic/Difference.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
+package org.scalactic
 
 trait Difference extends Serializable {
